% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnl_check_significance.R
\name{bnl_check_significance}
\alias{bnl_check_significance}
\title{bnl_check_significance}
\usage{
bnl_check_significance(
  score,
  lookup_table,
  num_cgs,
  quants_significance_cutoff = "0.999"
)
}
\description{
bnl_check_significance
}
