% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnl_rescore_wo_max_overlap.R
\name{bnl_rescore_wo_max_overlap}
\alias{bnl_rescore_wo_max_overlap}
\title{bnl_rescore_wo_max_overlap}
\usage{
bnl_rescore_wo_max_overlap(window, max_overlap, CG_table, FDR_scaler)
}
\description{
bnl_rescore_wo_max_overlap
}
