% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnl_select_max_scoring_overlap.R
\name{bnl_select_max_scoring_overlap}
\alias{bnl_select_max_scoring_overlap}
\title{bnl_select_max_scoring_overlap}
\usage{
bnl_select_max_scoring_overlap(overlaps, CG_table, FDR_scaler)
}
\description{
bnl_select_max_scoring_overlap
}
