% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_next_layer.R
\name{build_next_layer}
\alias{build_next_layer}
\title{build_next_layer}
\usage{
build_next_layer(
  prev_layer,
  windows_to_layer,
  CG_table,
  lookup_table,
  FDR_scaler
)
}
\description{
build_next_layer
}
