% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_significant_windows.R
\name{determine_significant_windows}
\alias{determine_significant_windows}
\title{determine_significant_windows}
\usage{
determine_significant_windows(
  window_results,
  indat,
  quants,
  quants_significance_cutoff = "0.9999"
)
}
\description{
determine_significant_windows
}
