% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmrscaler.R
\name{dmrscaler}
\alias{dmrscaler}
\title{dmrscaler}
\usage{
dmrscaler(
  locs,
  layer_sizes = c(4, 8, 16, 32, 64),
  layer_step_fracs = NA,
  fdrscaler,
  cltable,
  signif_cutoff = "0.9999"
)
}
\description{
dmrscaler
}
