% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_peak_significance.R
\name{estimate_peak_significance}
\alias{estimate_peak_significance}
\title{estimate_peak_significance}
\usage{
estimate_peak_significance(num_CG, score, lookup_table)
}
\description{
estimate_peak_significance
}
