% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_generate_dmr_tree.R
\name{example_generate_dmr_tree}
\alias{example_generate_dmr_tree}
\title{example_generate_dmr_tree}
\usage{
example_generate_dmr_tree(dmrscaler_result, layer, chr, start, stop)
}
\description{
example_generate_dmr_tree
}
