% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rand_index_matrix.R
\name{generate_rand_index_matrix}
\alias{generate_rand_index_matrix}
\title{Case Control Permutation Description Matrix}
\usage{
generate_rand_index_matrix(num_controls, num_cases, num_permutations)
}
\arguments{
\item{num_controls}{number of controls}

\item{num_cases}{number of cases}

\item{num_permutations}{number of permutations}
}
\description{
This function generates a matrix which describes
series of case-control label permutations.
}
