% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FDR_scalar.R
\name{get_FDR_scalar}
\alias{get_FDR_scalar}
\title{Get_FDR_scalar}
\usage{
get_FDR_scalar(MWW_FDR_table, MWW_FDR_threshold)
}
\arguments{
\item{MWW_FDR_table}{very specifically formated table == BAD DESIGN}

\item{MWW_FDR_threshold}{e.g. 0.2 for 20% FDR}
}
\description{
This function is a cludgy fix for a poorly designed FDR utility.
(NOTE THIS METHOD IS RATHER CRUDE, COULD STAND TO BE IMPROVED... SUGGESTIONS WELCOME :)
}
