% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loc_fdr_pval.R
\name{get_loc_fdr_pval}
\alias{get_loc_fdr_pval}
\title{get_loc_fdr_pval}
\usage{
get_loc_fdr_pval(
  mat,
  cases,
  controls,
  stat_test,
  fdr = 0.1,
  resolution = 100,
  return_table = FALSE
)
}
\arguments{
\item{mat}{matrix with samples on columns and locs on rows}

\item{cases}{case indices or column names}

\item{controls}{control indices or column names}

\item{stat_test}{statistical test used for signifance estimation}

\item{fdr}{false discover rate desired}

\item{resolution}{number of bins to use for table}

\item{return_table}{if set to TRUE returns table of FDR values instead of single value}
}
\value{
dmrscaler returns a list of dmrs at each layer
}
\description{
get_loc_fdr_pval
}
