% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_row_CpG.R
\name{initialize_row_CpG_df}
\alias{initialize_row_CpG_df}
\title{initialize_row_CpG_df}
\usage{
initialize_row_CpG_df(
  chr = 0,
  start_pos = 0,
  stop_pos = 0,
  numCG = 0,
  numCG_positive = 0,
  numCG_negative = 0,
  unsigned_bin_score = 0,
  pos_bin_score = 0,
  neg_bin_score = 0,
  unsigned_bin_sig = 0,
  pos_bin_sig = 0,
  neg_bin_sig = 0
)
}
\description{
initialize_row_CpG_df
}
