% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_single_chr_CpG_df.R
\name{initialize_single_chr_CpG_df}
\alias{initialize_single_chr_CpG_df}
\title{initialize_single_chr_CpG_df}
\usage{
initialize_single_chr_CpG_df(chr, num_spots)
}
\arguments{
\item{chr}{chromosome name}

\item{num_spots}{number of spots}
}
\description{
initialize_single_chr_CpG_df
}
