% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_nearest_window_scoring_func.R
\name{n_nearest_window_scoring_func}
\alias{n_nearest_window_scoring_func}
\title{n_nearest_window_scoring_func}
\usage{
n_nearest_window_scoring_func(indat, n_nearest, step_size, FDR_scaler)
}
\arguments{
\item{indat}{object that needs to be standardized and documented}

\item{n_nearest}{number of adjacent cgs to consider}

\item{step_size}{step fraction = how many windows cover each cg}

\item{FDR_scaler}{might not be important, controls type 1 error sort of}
}
\description{
n_nearest_window_scoring_func
}
