% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MWW.R
\name{run_MWW}
\alias{run_MWW}
\title{Run Mann Whitney Test}
\usage{
run_MWW(control_indices, case_indices, Beta)
}
\arguments{
\item{control_indices}{indices of controls in Beta matrix}

\item{case_indices}{indices of cases in Beta matrix}

\item{Beta}{matrix of beta values, proportion methylated, with rows as CGs and cols as samples}
}
\description{
This function runs the Mann Whitney U test
to generate significance values for difference
between case and control groups for each CG in the
Beta matrix. (SHOULD DEFINE HOW TO HANDLE TIES)
}
