% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MWW_rand_permutation.R
\name{run_MWW_rand_permutation}
\alias{run_MWW_rand_permutation}
\title{Run Case Control Permutation}
\usage{
run_MWW_rand_permutation(index_matrix, Beta, num_permutations)
}
\arguments{
\item{index_matrix}{matrix of permuted case-control labels}

\item{Beta}{matrix of methylation proportions unmethylated/methylated}

\item{num_permutations}{number of permutations}
}
\description{
This function runs the a series of
case-control label permutations. Recording
the result from each run
}
