% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_CpG_bin.R
\name{Score_CpG_bin}
\alias{Score_CpG_bin}
\title{Score_CpG_bin}
\usage{
Score_CpG_bin(numCG, CG_scores, FDR_scaler = 1, numCG_cutoff = 1)
}
\arguments{
\item{CG_scores}{vector of scores e.g. -log10pvalue}

\item{FDR_scaler}{score at which FDR is achieved}
}
\description{
This function is a cludgy fix for a poorly designed FDR utility.
(NOTE THIS METHOD IS RATHER CRUDE, COULD STAND TO BE IMPROVED... SUGGESTIONS WELCOME :)
}
