% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_CLT_lookup_table.R
\name{write_CLT_lookup_table}
\alias{write_CLT_lookup_table}
\title{Write CLT lookup table}
\usage{
write_CLT_lookup_table(
  num_reps,
  data_to_sample,
  FDR_scaler,
  clt_numCGs = c(2, 5, 10, 25, 50, 100, 150, 200, 250, 1000),
  quants = quants <- data.frame(row.names = c(0, 0.5, 0.75, 0.9, 0.95, 0.99, 0.999,
    0.9995, 0.9999, 0.99995, 0.99999, 1))
)
}
\description{
This function generates a lookup table
to estimate the likelihood that a window
represents a DMR feature (i.e. appears enriched in differntially methylated CGs)
(NOTE THIS METHOD IS RATHER CRUDE, SHOULD BE IMPROVED OR REPLACE)
}
