% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FDR_table.R
\name{write_FDR_table}
\alias{write_FDR_table}
\title{Write FDR Table}
\usage{
write_FDR_table(
  real_table,
  rand_table,
  FDR_bin_count = 50,
  FDR_table_row_names = "MWW="
)
}
\description{
This function uses the permutation and real data partition to
estimate the leve at which the desired FDR is achieved.
(NOTE THIS PART OF THE METHOD IS SUBOPTIMAL, SHOULD EVENTUALLY IMPROVE... SUGGESTIONS WELCOME :)
}
